LPS-83389 portal-search-elasticsearch6-impl: Move logic to DefaultSortTranslator
https://issues.liferay.com/browse/LPS-83389
